Initial changes to Common and Channel
