+++ AppInsightsCommon/.npmignore
@@ -1,0 +1,8 @@
+# NPM Ignore
+
+# ignore everything
+*
+
+# ... but these files
+!/package.json
+!/bundle/aicommon.*
