+++ AppInsightsCommon/Constants.ts
@@ -1,0 +1,2 @@
+///<summary>Function that returns property name which will identify that monitoring for given instance of XmlHttpRequest is disabled</summary>
+export var DisabledPropertyName: string = "Microsoft_ApplicationInsights_BypassAjaxInstrumentation";
