+++ gruntfile.js
@@ -18,8 +18,19 @@ module.exports = function (grunt) {
                     'coreSDK/JavaScriptSDK/*.ts',
                 ],
                 out: 'bundle/aicore.js',
-            }
-			,
+            },
+            common: {
+                src: [
+                    'AppInsightsCommon/*.ts',
+                ],
+                out: 'AppInsightsCommon/bundle/aicommon.js'
+            },
+            channel: {
+                src: [
+                    'AppInsightsChannel/*.ts',
+                ],
+                out: 'AppInsightsChannel/bundle/aichannel.js'
+            },
             module: {
                 src: [
                     'JavaScript/JavaScriptSDK.Interfaces/*.ts',
@@ -129,7 +140,9 @@ module.exports = function (grunt) {
     grunt.loadNpmTasks('grunt-contrib-qunit');
 
     grunt.registerTask("default", ["ts:default", "uglify:ai", "uglify:snippet"]);
-	grunt.registerTask("core", ["ts:core"]);
+    grunt.registerTask("core", ["ts:core"]);
+    grunt.registerTask("common", ["ts:common"]);
+    grunt.registerTask("channel", ["ts:channel"]);
     grunt.registerTask("module", ["ts:module"]);
     grunt.registerTask("test", ["ts:default", "ts:test", "ts:testSchema", "ts:testE2E", "ts:types", "qunit"]);
 };
